# aurora-go/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_aurora-go_aurora_internal_vo.ResourceVO:
    properties:
      id:
        description: 更新时必填，创建时为空
        type: integer
      isAnonymous:
        description: 是否匿名 0否1是
        type: integer
      parentId:
        description: 父模块ID，模块为null
        type: integer
      requestMethod:
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        - PATCH
        type: string
      resourceName:
        maxLength: 50
        type: string
      url:
        description: 模块不需要URL
        maxLength: 255
        type: string
    required:
    - resourceName
    type: object
...
host: localhost:8080
info:
  contact:
    email: aurora@example.com
    name: Aurora Team
  description: Aurora 博客系统 Go 1.26 后端 - 基于 tRPC-Agent-Go 的 AI 驱动博客平台
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/nicepkg/aurora
  title: Aurora Blog API
  version: "1.0"
paths:
...
  /api/admin/photos:
    delete:
      consumes:
      - application/json
      description: 后台批量删除照片
      parameters:
      - description: 照片ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 批量删除照片
      tags:
      - 照片
    get:
      consumes:
      - application/json
      description: 后台获取照片列表（分页）
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 相册ID
        in: query
        name: albumId
        type: integer
      - description: 当前页码
        in: query
        name: current
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回照片列表
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 后台照片管理列表
      tags:
      - 照片
    post:
      consumes:
      - application/json
      description: 保存照片到指定相册
      parameters:
      - description: 相册ID和照片URL列表(albumId, photoUrls)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 保存照片
      tags:
      - 照片
    put:
      consumes:
      - application/json
      description: 更新照片名称等信息
      parameters:
      - description: 照片ID和名称(id, photoName)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 更新照片信息
      tags:
      - 照片
  /api/admin/photos/album:
    put:
      consumes:
      - application/json
      description: 批量移动照片到指定相册
      parameters:
      - description: 目标相册ID和照片ID列表(albumId, photoIds)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 移动照片到其他相册
      tags:
      - 照片
...
  /api/admin/photos/delete:
    put:
      consumes:
      - application/json
      description: 批量更新照片删除状态（逻辑删除/恢复）
      parameters:
      - description: 照片ID列表和删除状态(ids, isDelete)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 更新照片删除状态
      tags:
      - 照片
  /api/admin/photos/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传照片文件
      parameters:
      - description: 照片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回照片URL
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 上传照片
      tags:
      - 照片
  /api/admin/resources:
    delete:
      consumes:
      - application/json
      description: 批量删除资源权限
      parameters:
      - description: 资源ID列表，逗号分隔
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 批量删除资源
      tags:
      - 资源
    get:
      consumes:
      - application/json
      description: 获取资源权限列表（树形结构）
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回资源树
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取资源权限列表
      tags:
      - 资源
    post:
      consumes:
      - application/json
      description: 后台保存或更新资源权限
      parameters:
      - description: 资源信息
        in: body
        name: resourceVO
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_vo.ResourceVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 保存/更新资源
      tags:
      - 资源
  /api/admin/resources/{id}:
    delete:
      consumes:
      - application/json
      description: 删除单个资源权限
      parameters:
      - description: 资源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 删除单个资源
      tags:
      - 资源
...
  /api/admin/role/resources:
    get:
      consumes:
      - application/json
      description: 获取角色授权资源选项（用于下拉框，树形结构）
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回资源树
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取角色资源选项
      tags:
      - 资源
...
  /api/admin/roles/{id}/resources:
    put:
      consumes:
      - application/json
      description: 为角色分配资源权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 资源ID列表(resourceIds)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 更新角色的资源关联
      tags:
      - 资源
...
  /api/albums/{albumId}/photos:
    get:
      consumes:
      - application/json
      description: 根据相册ID获取照片列表
      parameters:
      - description: 相册ID
        in: path
        name: albumId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回照片列表
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 根据相册ID获取照片列表
      tags:
      - 照片
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Token 认证，格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"